refactored moderately
